+++ goldens/circular-deps/packages.json
@@ -1129,6 +1129,10 @@
     "packages/forms/src/model.ts"
   ],
   [
+    "packages/forms/src/model.ts",
+    "packages/forms/src/utils/validators.ts"
+  ],
+  [
     "packages/language-service/src/completions.ts",
     "packages/language-service/src/template.ts",
     "packages/language-service/src/typescript_host.ts",
